app/pyproject.toml
[tool.poetry]
name = "flask-api"
version = "0.1.0"
description = ""
authors = ["Gatalist <kostencko.alexander2012@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.11"
flask = "2.0.2"
flask-login = "0.5.0"
flask-migrate = "3.1.0"
flask-sqlalchemy = "2.5.1"
flask-wtf = "1.0.0"
flask-admin = "1.6.0"
email-validator = "1.1.3"
psycopg2-binary = "2.9.6"
flask-bcrypt = "1.0.1"
flask-security = "3.0.0"
pytz = "2023.3"
gunicorn = "20.1.0"
werkzeug = "2.0"
flask-marshmallow = "^0.15.0"
flask-apispec = "^0.11.4"
sqlalchemy = "1.4.40"
marshmallow-sqlalchemy = "^0.29.0"
flask-restful = "^0.3.10"
loguru = "^0.7.0"
python-dotenv = "^1.0.0"
flask-httpauth = "^4.8.0"


[tool.poetry.group.dev.dependencies]
pylint = "2.15.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
